move list group functions to correct file
