+++ cognito/mock/group_mocker.go
@@ -3,10 +3,7 @@ package mock
 import (
 	"errors"
 	"fmt"
-	"math/rand"
 	"time"
-
-	"github.com/aws/aws-sdk-go/service/cognitoidentityprovider"
 )
 
 type Group struct {
@@ -23,34 +20,6 @@ func (m *CognitoIdentityProviderClientStub) AddGroupWithName(name string) error 
 		return err
 	}
 	m.Groups = append(m.Groups, newGroup)
-	return nil
-}
-
-func (m *CognitoIdentityProviderClientStub) AddGroupWithNameAndDescription(name, description string) error {
-	newGroup, err := m.GenerateGroup(name, description, 0)
-	if err != nil {
-		return err
-	}
-	m.Groups = append(m.Groups, newGroup)
-
-	pres := int64(0)
-	newGroupType := cognitoidentityprovider.GroupType{
-		CreationDate:     nil,
-		Description:      &description,
-		GroupName:        &name,
-		LastModifiedDate: nil,
-		Precedence:       &pres,
-		RoleArn:          nil,
-		UserPoolId:       nil,
-	}
-	newGroupTypeList := []*cognitoidentityprovider.GroupType{}
-	newGroupTypeList = append(newGroupTypeList, &newGroupType)
-	groupsListOutput := cognitoidentityprovider.ListGroupsOutput{
-		Groups:    newGroupTypeList,
-		NextToken: nil,
-	}
-	m.GroupsList = append(m.GroupsList, groupsListOutput)
-
 	return nil
 }
 
@@ -132,39 +101,4 @@ func (m *CognitoIdentityProviderClientStub) BulkGenerateGroups(groupCount int) {
 		m.Groups = append(m.Groups, &group)
 
 	}
-}
-
-func (m *CognitoIdentityProviderClientStub) BulkGenerateGroupsList(groupCount int) {
-	//Type to map for the Cognito GroupType object
-
-	var (
-		groupsList  cognitoidentityprovider.ListGroupsOutput
-		next_token  string = "next_token"
-		totalgroups        = 0
-	)
-
-	for i := 1; i <= groupCount; i++ {
-		D := "group name description " + fmt.Sprint(i)
-		G := "group_name_" + fmt.Sprint(i)
-		P := rand.Int63n(100-13) + 13
-
-		group := cognitoidentityprovider.GroupType{
-			Description: &D,
-			GroupName:   &G,
-			Precedence:  &P,
-		}
-
-		groupsList.Groups = append(groupsList.Groups, &group)
-
-		if i%60 == 0 && i > 0 {
-			groupsList.NextToken = &next_token
-			totalgroups = totalgroups + len(groupsList.Groups)
-			m.GroupsList = append(m.GroupsList, groupsList)
-			groupsList = *new(cognitoidentityprovider.ListGroupsOutput)
-		}
-
-	}
-	groupsList.NextToken = nil
-	m.GroupsList = append(m.GroupsList, groupsList)
-
 }
